requirements.txt
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
exceptiongroup==1.2.1
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
google-analytics==0.0.0
google-analytics-data==0.18.8
google-api-core==2.19.0
google-auth==2.30.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
selenium==4.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sob==1.50.8
sortedcontainers==2.4.0
soupsieve==2.5
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
